Split continues — explorer & profiler tabs moved to core/tabs/ .
Continues the split. ExplorerTab (1,239 lines) and ProfilerTab (476 lines) are
relocated into core/tabs/explorer_tab.py and profiler_tab.py. Pure move: classes
keep their methods; the main file imports them when building the Notebook.
Monolith: 22,474 → 20,773 lines. Per-module refactoring is deferred.
